requirements.txt
openai~=1.51.2
python-dotenv~=1.0.1
APScheduler~=3.10.4
peewee~=3.17.6
requests~=2.32.3
beautifulsoup4~=4.12.3
python-telegram-bot~=20.3